Rename(MemberList): 컴포넌트 폴더로 파일 이동
